refactor(backup): extract archive format, runtime bindings, paths, plans, records and export into a backend/core/backup package, taking backup_manager.py from 2875 lines back under its 2147-line grandfathered baseline so the BE-008 file-size gate passes again, introduce runtime.py as the single documented seam holding every binding the test harness substitutes (the surrogate models exist because the real ones use Postgres-only column types) so submodules resolve them late instead of binding at import and silently ignoring the patch, split the audio, document and config members out of _create_backup_sync and the per-table row rewrites out of _serialise_backup_table_rows to bring both back under the complexity gate rather than inheriting the old file's exemption, and bundle the export parameters into an _ExportRequest instead of an eight-argument signature
