+++ backend/core/backup/__init__.py
@@ -1,0 +1,1 @@
+"""Backup and restore for a whole Nojoin installation."""
